+++ packages/slidev/node/vite/loaders.ts
@@ -1,10 +1,10 @@
-import type { Connect, Plugin, ViteDevServer } from 'vite'
+import type { Plugin, ViteDevServer } from 'vite'
 import { notNullish, range } from '@antfu/utils'
 import type { ResolvedSlidevOptions, SlideInfo, SlidePatch, SlidevServerOptions } from '@slidev/types'
 import * as parser from '@slidev/parser/fs'
 import equal from 'fast-deep-equal'
 import type { LoadResult } from 'rollup'
-import { updateFrontmatterPatch } from '../utils'
+import { getBodyJson, updateFrontmatterPatch } from '../utils'
 import { templates } from '../virtual'
 import { templateTitleRendererMd } from '../virtual/titles'
 import { templateSlides } from '../virtual/slides'
@@ -14,22 +14,6 @@ import { templateMonacoTypes } from '../virtual/monaco-types'
 import { sharedMd } from '../commands/shared'
 import { createDataUtils } from '../options'
 import { regexSlideFacadeId, regexSlideReqPath, regexSlideSourceId } from './common'
-
-export function getBodyJson(req: Connect.IncomingMessage) {
-  return new Promise<any>((resolve, reject) => {
-    let body = ''
-    req.on('data', chunk => body += chunk)
-    req.on('error', reject)
-    req.on('end', () => {
-      try {
-        resolve(JSON.parse(body) || {})
-      }
-      catch (e) {
-        reject(e)
-      }
-    })
-  })
-}
 
 function renderNote(text: string = '') {
   let clickCount = 0
@@ -254,7 +238,7 @@ export function createSlidesLoader(
       const template = templates.find(i => i.id === id)
       if (template) {
         return {
-          code: await template.getContent(options, this),
+          code: await template.getContent.call(this, options),
           map: { mappings: '' },
         }
       }
